Port Changes to Setup and Example Files (#50)
* Updated main files

* Changed old exmaple files to benchmark and added new simplified EmbedDB example

* Renamed old examples to follow benchmark convention

* Continued reorganizing exmaple and benchmark files

* Started modifying example to be more customizable and to use specific sizes of integers

* Finished changing int types

* Identified issue with example file

* Formatted code
