+++ src/memboardMain.cpp
@@ -60,14 +60,21 @@ static ArduinoOutStream cout(Serial);
 #include "sdcard_c_iface.h"
 #include "serial_c_iface.h"
 
-#define EXAMPLE 2
+/**
+ * 0 - 2 are for benchmarks
+ * 3 is for the example program
+ *
+ */
+#define WHICH_PROGRAM 3
 
-#if EXAMPLE == 0
+#if WHICH_PROGRAM == 0
+#include "benchmarks/sequentialDataBenchmark.h"
+#elif WHICH_PROGRAM == 1
+#include "benchmarks/variableDataBenchmark.h"
+#elif WHICH_PROGRAM == 2
+#include "benchmarks/queryInterfaceBenchmark.h"
+#elif WHICH_PROGRAM == 3
 #include "embedDBExample.h"
-#elif EXAMPLE == 1
-#include "embedDBVariableDataExample.h"
-#elif EXAMPLE == 2
-#include "advancedQueryInterfaceExample.h"
 #endif
 
 #define ENABLE_DEDICATED_SPI 1
@@ -131,12 +138,14 @@ void setup() {
 
     init_df((void *)&at45db32_m);
 
-#if EXAMPLE == 0
+#if WHICH_PROGRAM == 0
     runalltests_embedDB();
-#elif EXAMPLE == 1
+#elif WHICH_PROGRAM == 1
     test_vardata();
-#elif EXAMPLE == 2
+#elif WHICH_PROGRAM == 2
     advancedQueryExample();
+#elif WHICH_PROGRAM == 3
+    embedDBExample();
 #endif
 }
